I moved the commented out classes to the old folders in 4 different packages for ease of use.
