+++ src/main/java/com/gmail/merikbest2015/ecommerce/controller/old/PerfumeController.java
@@ -1,4 +1,4 @@
-package com.gmail.merikbest2015.ecommerce.controller;
+package com.gmail.merikbest2015.ecommerce.controller.old;
 
 //import com.gmail.merikbest2015.ecommerce.mapper.PerfumeMapper;
 //import com.gmail.merikbest2015.ecommerce.service.graphql.GraphQLProvider;
