+++ src/main/java/com/gmail/merikbest2015/ecommerce/service/old/ReviewService.java
@@ -1,4 +1,4 @@
-package com.gmail.merikbest2015.ecommerce.service;
+package com.gmail.merikbest2015.ecommerce.service.old;
 
 //import com.gmail.merikbest2015.ecommerce.domain.Review;
 //
